Pull github interaction out of `http` module into a `github` module
